fix: tests; rename TimesWatermark
